fix: fix the bug that occurred allocate diary media to empty list - modify diary media save logic to add all, not allocate - moved handler and dto to diary
