+++ petlog-api/src/main/java/com/ppp/api/diary/dto/event/DiaryDeletedEvent.java
@@ -1,4 +1,4 @@
-package com.ppp.api.handler.dto;
+package com.ppp.api.diary.dto.event;
 
 import com.ppp.domain.diary.DiaryMedia;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class DiaryUpdatedEvent {
+public class DiaryDeletedEvent {
     private long diaryId;
     private List<DiaryMedia> diaryMedias;
 }
